+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- localization방식 LocalizationsDelegate에서 GetX 로 변경
+- get 3.4.1 -> 3.17.1
+- BarcodeDetail 별도 pakcage로 분리(BarcodeInfo)
 
 ## 1.6.0 - 2020-10-11
 ### Added